requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
python-multipart
asyncpg
sqlalchemy[asyncio]
fastapi-cache2 
redis
